changes and improvements
