+++ hooks/store/tags.tsx
@@ -1,6 +1,4 @@
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import toast from "react-hot-toast";
-import { useTranslation } from "next-i18next";
 import { TagIncludingLinkCount } from "@/types/global";
 
 const useTags = () => {
@@ -18,12 +16,9 @@ const useTags = () => {
 
 const useUpdateTag = () => {
   const queryClient = useQueryClient();
-  const { t } = useTranslation();
 
   return useMutation({
     mutationFn: async (tag: TagIncludingLinkCount) => {
-      const load = toast.loading(t("applying_changes"));
-
       const response = await fetch(`/api/v1/tags/${tag.id}`, {
         body: JSON.stringify(tag),
         headers: {
@@ -35,8 +30,6 @@ const useUpdateTag = () => {
       const data = await response.json();
       if (!response.ok) throw new Error(data.response);
 
-      toast.dismiss(load);
-
       return data.response;
     },
     onSuccess: (data) => {
@@ -45,22 +38,15 @@ const useUpdateTag = () => {
           tag.id === data.id ? data : tag
         )
       );
-      toast.success(t("tag_renamed"));
-    },
-    onError: (error) => {
-      toast.error(error.message);
     },
   });
 };
 
 const useRemoveTag = () => {
   const queryClient = useQueryClient();
-  const { t } = useTranslation();
 
   return useMutation({
     mutationFn: async (tagId: number) => {
-      const load = toast.loading(t("applying_changes"));
-
       const response = await fetch(`/api/v1/tags/${tagId}`, {
         method: "DELETE",
       });
@@ -68,18 +54,12 @@ const useRemoveTag = () => {
       const data = await response.json();
       if (!response.ok) throw new Error(data.response);
 
-      toast.dismiss(load);
-
       return data.response;
     },
     onSuccess: (data, variables) => {
       queryClient.setQueryData(["tags"], (oldData: any) =>
         oldData.filter((tag: TagIncludingLinkCount) => tag.id !== variables)
       );
-      toast.success(t("tag_deleted"));
-    },
-    onError: (error) => {
-      toast.error(error.message);
     },
   });
 };
